+++ contracts/Core/storage/StakeStorage.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.0;
 import "../../lib/Structs.sol";
 
 contract StakeStorage {
-    uint256 public numStakers;
+    uint32 public numStakers;
 
-    mapping(address => uint256) public stakerIds;
-    mapping(uint256 => Structs.Staker) public stakers;
+    mapping(address => uint32) public stakerIds;
+    mapping(uint32 => Structs.Staker) public stakers;
     mapping(address => mapping(address => Structs.Lock)) public locks;
 }
